pkg restructuring
