+++ internal/cli/completion.go
@@ -1,4 +1,4 @@
-package commands
+package cli
 
 import (
 	"os"
